requirements.txt
tensorflow
streamlit
numpy
pandas
matplotlib
scikit-learn
pillow
h5py
